Implemented path based correlation functions.
