+++ checker/src/main/kotlin/com/dickow/chortlin/checker/correlation/path/Path.kt
@@ -1,5 +1,7 @@
 package com.dickow.chortlin.checker.correlation.path
 
+import com.dickow.chortlin.checker.trace.value.Value
+
 interface Path {
-    fun apply(input: Any?): Any
+    fun apply(input: Value): Any
 }
