+++ checker/src/main/kotlin/com/dickow/chortlin/checker/trace/value/BooleanValue.kt
@@ -1,0 +1,3 @@
+package com.dickow.chortlin.checker.trace.value
+
+data class BooleanValue(val boolean: Boolean) : Value
