+++ checker/src/main/kotlin/com/dickow/chortlin/checker/trace/value/ListValue.kt
@@ -1,0 +1,3 @@
+package com.dickow.chortlin.checker.trace.value
+
+data class ListValue(val valueList: List<Value>) : Value
