+++ checker/src/main/kotlin/com/dickow/chortlin/checker/trace/value/NumberValue.kt
@@ -1,0 +1,3 @@
+package com.dickow.chortlin.checker.trace.value
+
+data class NumberValue(val number: Number) : Value
