+++ checker/src/main/kotlin/com/dickow/chortlin/checker/trace/value/ObjectValue.kt
@@ -1,0 +1,3 @@
+package com.dickow.chortlin.checker.trace.value
+
+data class ObjectValue(val values: Map<String, Value>) : Value
