+++ checker/src/main/kotlin/com/dickow/chortlin/checker/trace/value/RootValue.kt
@@ -1,0 +1,3 @@
+package com.dickow.chortlin.checker.trace.value
+
+data class RootValue(val value: Value) : Value
