+++ checker/src/main/kotlin/com/dickow/chortlin/checker/trace/value/StringValue.kt
@@ -1,0 +1,3 @@
+package com.dickow.chortlin.checker.trace.value
+
+data class StringValue(val string: String) : Value
