+++ checker/src/main/kotlin/com/dickow/chortlin/checker/trace/value/Value.kt
@@ -1,0 +1,3 @@
+package com.dickow.chortlin.checker.trace.value
+
+interface Value
