--- test-application/src/test/kotlin/com/dickow/chortlin/core/test/shared/OnlineCheckerClasses.kt
@@ -1,31 +1,0 @@
-package com.dickow.chortlin.core.test.shared
-
-import com.dickow.chortlin.interception.annotations.TraceInvocation
-import com.dickow.chortlin.interception.annotations.TraceReturn
-
-class OnlineFirstClass {
-    @TraceInvocation
-    @TraceReturn
-    fun method1() {}
-    @TraceInvocation
-    @TraceReturn
-    fun method2() {}
-}
-
-class OnlineSecondClass {
-    @TraceInvocation
-    @TraceReturn
-    fun method1() {}
-    @TraceInvocation
-    @TraceReturn
-    fun method2() {}
-}
-
-class OnlineThirdClass {
-    @TraceInvocation
-    @TraceReturn
-    fun method1() {}
-    @TraceInvocation
-    @TraceReturn
-    fun method2() {}
-}
