--- test-application/src/test/kotlin/com/dickow/chortlin/core/test/shared/objects/Receipt.kt
@@ -1,16 +1,0 @@
-package com.dickow.chortlin.core.test.shared.objects
-
-class Receipt constructor() {
-    constructor(itemId: Int, price: Int, itemName: String, sellerId: String) : this() {
-        this.itemId = itemId
-        this.price = price
-        this.itemName = itemName
-        this.sellerId = sellerId
-    }
-
-    private var itemId: Int = 0
-    private var price: Int = 0
-    private lateinit var itemName: String
-    private lateinit var sellerId: String
-
-}
