+++ test-application/src/test/kotlin/com/dickow/chortlin/test/application/shared/OnlineCheckerClasses.kt
@@ -1,0 +1,31 @@
+package com.dickow.chortlin.test.application.shared
+
+import com.dickow.chortlin.interception.annotations.TraceInvocation
+import com.dickow.chortlin.interception.annotations.TraceReturn
+
+class OnlineFirstClass {
+    @TraceInvocation
+    @TraceReturn
+    fun method1() {}
+    @TraceInvocation
+    @TraceReturn
+    fun method2() {}
+}
+
+class OnlineSecondClass {
+    @TraceInvocation
+    @TraceReturn
+    fun method1() {}
+    @TraceInvocation
+    @TraceReturn
+    fun method2() {}
+}
+
+class OnlineThirdClass {
+    @TraceInvocation
+    @TraceReturn
+    fun method1() {}
+    @TraceInvocation
+    @TraceReturn
+    fun method2() {}
+}
